Sample Python/Flask app using the Uber api
The goal of the project is to provide a working example of the Uber
api. It is intended to be simple, easy to understand and a platform off
of which developers can build their own applications. For more
information about how to use this example, please read the README.md
file for more information and for instructions for how to run this
project.